pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "signal-replay"
version = "1.0.0"
description = "Python package for replaying high-resolution event logs from ATC signal controllers back to test controllers using NTCIP"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "ATC Signal Replay Contributors"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
keywords = ["traffic", "signal", "controller", "NTCIP", "SNMP", "replay", "simulation"]

dependencies = [
    "atspm",
    "pysnmplib>=5.0.24",
    "dtaidistance",
    "matplotlib",
    "pyyaml"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/your-org/ATC-Signal-Replay"
Documentation = "https://github.com/your-org/ATC-Signal-Replay#readme"
Repository = "https://github.com/your-org/ATC-Signal-Replay"
Issues = "https://github.com/your-org/ATC-Signal-Replay/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
signal_replay = ["sql/*.sql"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]
